pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "TextArena"
version = "0.5.9"
description = "[WIP] A Collection of Competitive Text-Based Games for Language Model Evaluation and Reinforcement Learning"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Leon Guertler", email = "Guertlerlo@cfar.a-star.edu.sg" }]
dependencies = [
    "requests",
    "openai",
    "rich",
    "nltk",
    "chess",
    "networkx",
    "python-dotenv",
    "opencv-python",
    "websockets",
    "mss",
    "mcp",
    "anthropic",
]

[project.urls]
"Homepage" = "https://github.com/LeonGuertler/TextArena"
